Put tests in single file
