+++ jest.config.cjs
@@ -1,6 +1,7 @@
 module.exports = {
   "roots": [
-    "<rootDir>/src"
+    "<rootDir>/src",
+    "<rootDir>/test",
   ],
   "testMatch": [
     "**/__tests__/**/*.+(ts|tsx|js)",
